Add support for a 'binary main()' function.
In an environment such as CloudABI we want program arguments to be more
powerful than before. Instead of just providing command line arguments
and environment arguments, it may be more useful to provide a single
binary buffer instead. This buffer can use a format such as JSON, XML,
YAML or Protocol Buffers to provide parameters hierarchically.

Add a fallback program_main() function to the C library that calls into
main(). This keeps existing applications working, but still provides the
framework for programs that want binary data.