Extension: add fswap().
Libraries like libc++ assume that the stderr object never changes its
address. This makes it impossible for us to make cerr work, as by
default, stderr simply discards data.

Add a new function called fswap(). This function can be used to exchange
the contents of two FILE objects in a thread-safe manner. This function
can be used to alter the output of stderr at any point in time.