Heavily refactor crt0.
Add a new function mi_startup() that we can use for all
machine-independent code. It parses the ELF program header, sets up TLS
and patches up the global variables.

The _startup() function then becomes machine dependent. For x86-64 it is
a complete mess, as we still need to fix up the kernel to pass arguments
to the process in a saner way.