Make it possible to convert stock tzdata files.
The gen_tzdata script didn't support translation of rules that used
'last${weekday}' in combination with 'Feb', spanning both leap and
non-leap years. It should either translate it to 'Day>=22' or 'Day>=23'.

It turns out that for our dataset, it is always safe to return
'Day>=22', as there is not a single entry for which we ever had a
transition on February 29th. Make the script emit 'Day>=22'
unconditionally.