First cut at adding support for argument data.
The raw exec() interface exposed by the kernel allows us to provide a
buffer of data and a list of file descriptors to the new process. This
is far too low-level.

Introduce argdata_t, which behaves as a JSON-esque data structure. File
descriptors are encoded as a separate type -- not as an integer. The
advantage of this approach is that it allows functions like exec() and
pdexec() to extract which file descriptors have to be passed to the new
process.