Provide a copy of nl_langinfo() that doesn't depend on malloc(), etc.
nl_langinfo_l() is a pretty heavy function, as it possibly needs to
translate wide character/UTF-8 strings to ASCII. It needs to store the
results using malloc(), etc.

Just provide a simple copy that returns the same values as
nl_langinfo(LC_C_LOCALE). The unit tests already check for this.