Add a rudimentary ungetc() implementation.
The nice thing is that it is implemented entirely within the top layer
of stdio. We add a separate buffer to the FILE object that can hold up
to four wide characters. When we call fread_peek(), we return this
buffer instead of the underlying read buffer when there are pending
characters.

The tests are extended to test ungetc() as well. For now we only test
fgetc()/getc(). It would be nice if we could also test
fread()/fgets()/etc.