Remove the open flags from stdio objects.
In an attempt to reduce the number of invariants inside of the stdio
layer, let's remove the oflags field entirely. If a stream is opened for
reading, simply present an alternative file operations structure that
has write_peek() set to a function that returns EBADF.

This way fread_{peek,consume}() and fwrite_put() become simpler. The
machine code for the top half of stdio is now also smaller.