Add stub for POSIX message queues.
Though we won't be able to implement the POSIX message queue global
namespace (mq_open and mq_unlink), it should be possible to provide
support for anonymous message queues.

Let's start off by implementing a simple message queue that is only
process-local. If we need to port applications that require
cross-process message queues, we will extend this to be, say, shared
memory backed.