Completely implement mq_receive().
I was mistaken; we can't use a binary heap in this situation as we still
need to look up the priority later on to add new messages.

We could still use a binary search tree for this purpose, but I think
we're over-engineering things if we're going in that direction. For now,
just use a simple linked list of priorities. This means that extracting
the message still runs in constant time.