Add a test for the ordering of messages stored in the queue.
This test stores 4096 one-byte messages in the queue, split up across 32
different priority levels. Messages of different priorities are
interleaved a random. We should be able to fetch them back from the
message queue in sorted order.

This allowed us to already detect two bugs in the implementation. One
where we didn't initialize the message priority, and another one where
we traversed the linked list across the wrong pointer.