Replace tsearch() by a new implementation.
It turns out that there is no need to explicitly keep track of the
height of subtrees. Just storing the balance is enough. This new
implementation is therefore more efficient. There is no need to
recompute any subtree balances.

It now also becomes more easy to determine when we can just stop
mutating the tree. It may well be the case that the insertion of a leaf
only causes changes to some of the nodes at the bottom of the tree. We
now stop the recursion more quickly.

tdelete() still needs to be adjusted.