Replace tsearch() and tdelete() by iterative implementations.
Undo the recursion by manually keeping track of an array of parent
pointers. We know that this array never needs to be larger than
1.5*bits_in_a_pointer entries. This approach has a couple of advantages:

- Memory usage is now fixed to an exact bound.
- We can immediately return in case we find a matching entry in
  tsearch() (or find no matching entry in tdelete()).
- We can immediately return as soon as we don't need to do any
  rebalancing.

The machine code generated is also smaller than for the recursive
functions. tsearch() stays approximately the same size, whereas
tdelete() is now 20% smaller, meaning it is not much larger than
tsearch().