Reimplement tsearch() and tdelete() to use O(log(n)) space.
The current implementation uses O(log(n)^2) space, due to the fact that
the parents array is comprised of pointers. This means we have to
reserve 768 bytes on 64-bit systems just to search through the tree.

In this new implementation we only need 96 bits, but use 128. This is
accomplished by simply walking forward along the tree twice. We only
need a single bit to store whether we should go left or right.

Unfortunately this introduces some duplication in the code responsible
for rebalancing, which I'm not happy about. Apart from that, all seems
to work fine.