Don't store the table entries in the hash table structure directly.
Doing so has the disadvantage that our hash table is always slightly
larger than a power of two. As our implementation of malloc() uses power
of two buckets (up to a certain size), we currently waste quite a lot of
space as a result of this.

Simply let the hash table structure point to the dynamically allocated
hash table.