Add support for signal sets.
Even though CloudABI does not support installing signal handlers, there
is no reason why an application cannot keep track of signals in a set in
a way it likes, for example for keeping track of the signals that caused
child processes to terminate.

Provide support for sigaddset(), sigdelset(), sigemptyset(),
sigfillset() and sigismember().