Test against the right error code.
After reading up on some more specifications and comparing
implementations, it looks like ECONNRESET is the right error code to
return. POSIX connect() describes that the function can return
ECONNRESET, not ECONNABORTED. This is essentially a connect() failing
asynchronously.

POSIX also further states that ECONNABORTED has to be returned if the
connection was aborted locally. In this case we're closing the other
side of the UNIX socket, so the failure is happening remotely.