Add LOG_UPTO().
Python seems to depend on the LOG_UPTO() macro that can be used to
construct logging priority bitmasks, including all priorities up to and
including a given priority. As Python uses it unconditionally, I assume
that it is available on virtually any system out there. Let's go ahead
and stay compatible with the rest.