Correct the definitions of S_TYPEIS*().
It turns out that these macros should take a pointer to the stat
structure that is tested. This is unlike the other macros that receive
the st_mode field.

By fixing these to work properly, we can no longer use them in static
assertions. This is likely not going to be a problem, because we already
have coverage for these macros in other places.