Add the usleep() function.
While trying to get memcached to build, I noticed that it uses the
usleep() in a couple of places. I thought this function was never part
of POSIX, but I was mistaken. It was only marked obsolete in the 2004
edition. Let's stop being overly pedantic and add this function.

POSIX doesn't require implementations to accept microsecond values of
1000000 or higher, but I think we should spend no effort to restrict
this. Most other implementations also don't restrict this and I guess
this extension is the reason why it's still being used. It allows you to
sleep without the need to mess around with timeval/timespec structures.