Provide proper iterators for mappings/sequences.
The old API forced you to iterate through entries in mappings and
sequences by providing a callback. This not only has the disadvantage
that it typically requires you to write more code, it doesn't allow the
traversal to be suspended.

Introduce argdata_map_iterator_t and argdata_seq_iterator_t. These
objects keep track of the position with a mapping or sequence. They have
an 'error' field that can be used to accumulate error conditions. All
other members are hidden, as we don't want to export the innards of
argdata_t itself.

Patch up the existing consumers of the old interface to use the new one.

Initial patch by: Maurice Bos