Add support for Position Independent Executables to crt0.
If we want to support PIE, we need to extend crt0 to have a relocator.
We must also ensure that we don't depend on any global variables until
the relocator as finished executing. Therefore, make the top of _start()
store everything it has extracted from the auxiliary vector and program
headers in local variables. Write them back to the globals once
finished.

The relocator currently implements the relocations that I've observed on
x86-64, using Clang and GNU ld (BFD). LLD doesn't seem to work at all.
GNU ld also misbehaves on aarch64, as it discards R_AARCH64_ABS64
relocations. Do implement the relocations that did work properly on
aarch64.

While there, also rename the variables that we use to store the things
extracted from the auxiliary vector and program headers to be more
consistent.