Manually declare the system call table.
As discussed with Maurice, we want CloudABI executables to become
unaware of how to invoke system calls. This should entirely be solved by
passing in a vDSO into the application that contains a function for
every system call.

For now, allocate a system call table that is filled with the inline
assembly functions. We'll later replace this by a table that is filled
with ENOSYS functions.