Add simple const preservation to some of the standard library functions.
Functions like strchr() always have the downside that they silently
discard const qualifiers. Introduce a __preserve_const() macro that can
be used to wrap functions so that they propagate the const keyword to
the pointer type that is being returned.

In C++ this can normally be solved by using overloading. In C, let's
stick to using _Generic(). Though the macro is not 100% accurate, it
should never falsely return const.