Also add l64a_r().
As we're not going to add l64a(), we should still add an inverse
function of a64l(). It turns out that SVID has an l64a_r() that is
implemented by various systems (Solaris, HP-UX, the BSDs). Let's go
ahead and also add an implementation to our C library.