Rework the CMSG_*() functions for good.
It turns out that my previous approach wasn't good either. The API in
general effectively requires that the buffer provided in msg_control is
properly aligned. The msg_controllen field is allowed to be a sum of
CMSG_SPACE() values, meaning that its value must be tight.

Fall back to an approach similar to Linux and BSD, but do add a lot of
safety belts to the implementation to require that the cmsghdrs are
always properly aligned.

This API is such a mess.