Add mblen(), mbtowc() and wctomb().
It turns out that I was completely mistaken about the intent of these
functions. It looks like the only scenario in which they cannot be
implemented thread-safe is when we want to support shift-state character
sets, which we are not planning to.

Provide them, including _l() variants. This allows us to revert some
local changes we had to make to libc++.