Rework the way mbstate_t is implemented.
Instead of making mbstate_t an opaque object that can be implemented
differently by the wchar and uchar code, come up with a common format
that is both capable of storing surrogate halves and partially parsed
UTF-8 and GB 2312 characters.

Instead of letting code mess around with the innards of mbstate_t
directly, come up with a simple API that allows us to ensure that object
is always in a consistent state (read: not having both a surrogate half
and a partially parsed character). This makes it trivial to implement
mbsinit(), which we now finally support.