Revert "Add mbsinit_l()."
On second thought, let's not add this function. Current style of this
library seems to be to not add these kinds of functions. For example, we
also don't have putwc_l() and wcwidth_l(). *_l() functions are only
added if they are part of the standard or provide actual functionality.