Rework our workaround for Easter Island to make sense.
Applying DST twice just to get rule matching to work is madness. After
closer inspection, the problem is not related to daylight saving time.
It's just that switching to the next era causes the time to invisibly go
forward by an hour. This is the time value used for matching.