Properly deallocate temporary strings upon failure.
Let the function keep track of an array of pointers to objects that need
to be freed in case the function call fails. By calling
FREELIST_APPEND(), we can push objects into that list. Upon success,
only the freelist is deallocated. Otherwise, the objects that are
referred are deallocated as well.