Give sa_data a proper length.
It turns out that there is some code that uses sizeof(sockaddr:sa_data),
so that it can give a binary dump of the sockaddr's contents. This is
pretty useless in my opinion, but let's do at least something to make
such code build.

Declare sa_data as a zero-element array, so that the code at least
builds.