Make strptime() more complete.
Allow it to recombine week numbers and weekdays into full dates. Support
all three schemes supported by strftime() (first week starts on first
Sunday, first Monday or ISO 8601).

To test this, redo the unit tests for strptime() and add a randomized
component, where we parse random timestamps with a set of formats that
should parse back the entire date.

An advantage of this new implementation is that the structure returned
by strptime() is now always fully consistent. For example, only
providing the date, it is possible to use tm_wday and tm_yday.