Make handling of partial data complete.
There are some pieces of code that use strptime() to only parse the name
of a month or day. The latter didn't work, as the weekday was only used
as part of week number processing.

If a weekday is provided without a week number, simply compute the first
day of the year that matches.