Add implementations of SHA-256 and SHA-512.
In order to get Python's 'crypt' module working, we'll need a functional
implementation of crypt(). In our case we're going to pick the
thread-safe crypt_r() function instead.

At the core of this function, we'll need implementations of
cryptographically-safe hashing functions such as SHA-256 and SHA-512.
Import a copy of these functions that is almost a literal copy-paste
from the specification.

Code that can be shared between these two hashing functions is stored in
a separate source file. With some #defines, we ensure that this can be
compiled twice into two separate variants.