Add crypt_r().
The POSIX function for generating password hashes, crypt(), is not
thread-safe. This is the reason we're not implementing it.
Unfortunately, this means we're missing out on things like having a
working 'crypt' module for Python.

To alleviate this, add an implementation of crypt_r(), which is a
function that is also provided by glibc and musl. To keep our
implementation reasonably compact, only provide support for MD5, SHA-256
and SHA-512 hashes. Blowfish isn't used that frequently. Triple-DES is
pretty much outdated.