No longer depend on <cloudabi_syscalls_native.h>.
The invocation of hardware system calls is going to be outside of the
scope of the CloudABI spec, so we're not going to provide those anymore.
Only in the form of a vDSO that OS authors can decide to use (or not).

Copy a bunch of system calls that we need into syscall_fallback.h. At
some point in time we're going to remove it.