Disable the relocator on i386.
I'd say it makes most sense to disable PIE on i386. i386 doesn't have
any RIP-relative addressing, meaning that we have to apply a lot of
relocations to get things to work. Also, the relocations would need to
be done by the kernel, which is likely going to be out of the question.