Reimplement tm_zone to be more similar to Linux/FreeBSD.
Instead of storing the timezone abbreviation directly inside of the tm
structure, use a pointer instead. This is also what other systems like
Linux and FreeBSD do. It has the advantage that we can leave TZNAME_MAX
completely undefined.

Unfortunately, this means that localtime_l() needs to become a bit more
complex: we need to use a stringpool here to store cached copies of the
timezone abbreviations in case we need to create them dynamically.

This should allow us to remove a local change to Python.