Turn clockid_t into a pointer to an object.
POSIX allows clockid_t to be a pointer as well, so let's make use of
this fact by turning all of the individual clocks into objects. This has
the advantage that it becomes a lot easier to implement
pthread_getcpuclockid(), as these can simply be objects that are
embedded in the pthread_t that have their own logic.