Add faccessat().
Removing this system call seems to be a bit overly strict. In our case
there is some genuine use for this system call: it could be used to just
validate the file descriptor's rights.

This is nice about a capability-based system. This system call is
traditionally prone to TOCTOU problems, but in our environment it is
not. The rights are purely based on file descriptor state; not metadata
stored at the filesystem level.