Add backtrace_symbols() and backtrace_symbols_fd().
Now that we have a working copy of dladdr(), go ahead and add
implementations of the GNU extensions backtrace_symbols() and
backtrace_symbols_fd(). The idea is that we should also have a
backtrace() function. Once we have that, I'm going to extend functions
like __assertion_failed() to print stack traces.