Make thread deallocation in case of detaching work properly.
At first I wanted to implement this by creating queues of threads that
still needed to be cleaned up. Instead of doing this, simply store a
single thread in a global variable and call pthread_join() on it prior
to termination.

Though this doesn't scale with lots of threads, it looks like thread
detaching isn't used often enough to worry about this for now.

Reviewed by: Sjors Gielen