Implement unbuffered and line buffered streams properly.
Sjors Gielen reported that this aspect of stdio has not been implemented
properly. Extend fwrite_produce() to store in a boolean whether flushing
should be performed. In that case, funlockfile() can call into
fop_flush().

This approach has the advantage that the number of actual flushes
performed is minimised. For example, fprintf() can still be implemented
using putc_unlocked(), without causing it to do a flush for every single
character.