Remove msghdr::msg_name, getpeername(), getsockname() and recvfrom().
Socket addresses have the disadvantage that they cannot be
user-injected. When proxying/load balancing requests, it's impossible to
retain and propagate the original client/server addresses.

Now that we're working on Flower, there should be less of a need to
request this information from the kernel. Instead, programs should just
print/log the key-value pairs reported by Flower.

Go ahead and remove all APIs that normally report socket addresses. The
only exception is accept(), which still has to stay for obvious reasons.
Change that function to just set the sockaddr structure that is provided
to AF_UNSPEC.