Don't allow poll()'s events to be zero.
In this implementation, there is no way for us to poll for just POLLHUP,
POLLERR and POLLNVAL without specifying POLLRDNORM or POLLWRNORM. Every
event is decomposed into up to two separate subscriptions. If we would
create zero subscriptions, we couldn't listen for the error conditions.

Just return ENOSYS. Update the tests accordingly.