Add FIONBIO.
There seem to be various pieces of software that make use of this
ioctl(), even as part of their public API. Be a bit less pedantic and
provide it for now.

When we're going to do the file descriptor I/O refactoring, blocking
flags are going to be moved out of the kernel anyway. In that case it
will require even less code to have this in userspace.