Remove kqueue().
The kqueue() interface has the problem that it is fairly hard to
implement correctly, even though it isn't necessarily what we'd want.
Libraries like libuv demonstrate that what applications want is not
exactly what kqueue() offers. For example, there is no need to register
any timers. A simple timeout is sufficient. The kqueue() interface also
doesn't offer any support for asynchronous I/O, which is something
libraries like libuv would strongly benefit from.

Let's use this opportunity to go to the drawing board and come up with
something that is actually useful, easier to reason about and at least
as performant.