Import a custom version of libuv.
Libuv provides an event loop. Its API is pretty well thought out. In my
opinion it makes more sense to work towards having a library like this
working than to provide a non-standard polling framework.

This change brings in a custom version of libuv that tries to stick to
the official API. All interfaces incompatible with CloudABI's security
model are absent. The uv_process_t layer is implemented in such a way
that it uses process descriptors to start processes, using Argdata.