Remove pdfork() and everything related to it.
Though powerful, forking is known for causing lots of churn. In the case
of CloudABI, forking is also quite undesired, as it's rough to come up
with an API to properly separate capabilities.

Now that we're focussing on implementing libuv, we'll have a pretty nice
API for spawning processes. Let's just stick to that. Once the use of
that has become more common, we'll go ahead and remove pdwait() and
program_spawn() as well.