Properly implement uv_process_kill().
CloudABI doesn't have a separate kill system call. This is slightly
annoying, as it means there is no way to detect whether a process has
terminated or you killed it without any race conditions.

Still, let's add a uv_process_kill() that somewhat does what we expect,
namely closing the underlying process descriptor. This not only
increases portability, it may make code a bit cleaner, as observed when
porting Scuba to use this.