Add <arpa/nameser.h>.
This is a non-POSIX header file that exists on most UNIX-like systems,
containing various definitions specific to DNS. This header file is
normally used by c-ares. c-ares will fall back to an internal copy if
it's missing, but the annoying thing is that these definitions are used
as part of the c-ares API as well.

This header file contains the smallest number of definitions needed to
make c-ares and the Node.js c-ares bindings build.