Add stubs for uv_udp_*() functions we can support properly.
The official implementation of libuv requires that UDP sockets are not
connected. This is something that's not supported by CloudABI, because
it would allow us to send arbitrary UDP traffic.

In the long run I'd like to open the discussion with the libuv
developers about also supporting connected sockets. In the meantime, I'd
like it if we already implemented support for receiving UDP traffic.