Translate ENOTCAPABLE returned by proc_exec() to {EBADF,EACCES}.
FreeBSD's implementation of proc_exec() already returns EBADF and EACCES
in the appropriate cases. On a cleaner implementation of this
environment, it is fairly realistic that the system call returns
ENOTCAPABLE, due to it checking rights bits more accurately/earlier on.
Add proper translation of this error number to EBADF and EACCES to
remain POSIX-like.

Discussed with: Sjors Gielen