Rename program_getuuid() to program_getpid().
Now that the kernel emits CLOUDABI_AT_PID records containing a UUIDv4,
reimplement this function to make use of the info. While there, rename
it, as this name is slightly better: it's not just a UUID that we
return, it's an identifier for the process. It also makes it more
obvious it's a replacement for getpid().