Rename uv_spawn() to program_spawn().
The current name is annoying, as it is a name that is also used by libuv
officially, even though the semantics are different. It also makes it
harder to implement <program.h> on non-CloudABI systems to act as a
launcher for CloudABI processes.

Simply add this function to <program.h>, so it can be next to all of the
other CloudABI specific program management functions.