Migrate certain unit tests to Google Test.
Back when we started working on this library, we had to write our own
unit testing framework because C++ wasn't available yet. We also didn't
want to have any external dependencies, as that would complicate the
build procedure significantly, due to cyclic dependencies (cloudlibc ->
Google Test -> cloudlibc).

With Bazel, it is pretty easy to deal with this properly. Switch over to
a copy of Google Test that is provided by our bazel-third-party
repository. Already do this for a couple of tests to test feasibility.